+++ activerecord/test/cases/view_test.rb
@@ -1,7 +1,9 @@
 require "cases/helper"
 require "models/book"
+require "support/schema_dumping_helper"
 
 module ViewBehavior
+  include SchemaDumpingHelper
   extend ActiveSupport::Concern
 
   included do
@@ -31,6 +33,15 @@ module ViewBehavior
     assert_equal ["Ruby for Rails"], books.map(&:name)
   end
 
+  def test_views
+    assert_equal [Ebook.table_name], @connection.views
+  end
+
+  def test_view_exists
+    view_name = Ebook.table_name
+    assert @connection.view_exists?(view_name), "'#{view_name}' view should exist"
+  end
+
   def test_table_exists
     view_name = Ebook.table_name
     assert @connection.table_exists?(view_name), "'#{view_name}' table should exist"
@@ -53,6 +64,11 @@ module ViewBehavior
     end
     assert_nil model.primary_key
   end
+
+  def test_does_not_dump_view_as_table
+    schema = dump_table_schema "ebooks"
+    assert_no_match %r{create_table "ebooks"}, schema
+  end
 end
 
 if ActiveRecord::Base.connection.supports_views?
@@ -70,6 +86,7 @@ class ViewWithPrimaryKeyTest < ActiveRecord::TestCase
 end
 
 class ViewWithoutPrimaryKeyTest < ActiveRecord::TestCase
+  include SchemaDumpingHelper
   fixtures :books
 
   class Paperback < ActiveRecord::Base; end
@@ -91,6 +108,15 @@ class ViewWithoutPrimaryKeyTest < ActiveRecord::TestCase
     assert_equal ["Agile Web Development with Rails"], books.map(&:name)
   end
 
+  def test_views
+    assert_equal [Paperback.table_name], @connection.views
+  end
+
+  def test_view_exists
+    view_name = Paperback.table_name
+    assert @connection.view_exists?(view_name), "'#{view_name}' view should exist"
+  end
+
   def test_table_exists
     view_name = Paperback.table_name
     assert @connection.table_exists?(view_name), "'#{view_name}' table should exist"
@@ -108,6 +134,11 @@ class ViewWithoutPrimaryKeyTest < ActiveRecord::TestCase
 
   def test_does_not_have_a_primary_key
     assert_nil Paperback.primary_key
+  end
+
+  def test_does_not_dump_view_as_table
+    schema = dump_table_schema "paperbacks"
+    assert_no_match %r{create_table "paperbacks"}, schema
   end
 end